# docs/swagger.yaml
basePath: /
definitions:
...
  forms.CreateFormDto:
    properties:
      description:
        maxLength: 1000
        type: string
      title:
        maxLength: 255
        minLength: 2
        type: string
    required:
    - title
    type: object
  forms.FormDetailResponse:
    properties:
      form:
        $ref: '#/definitions/forms.FormResponse'
      sections:
        items:
          $ref: '#/definitions/forms.FormSectionWithFields'
        type: array
    type: object
  forms.FormFieldDetail:
    properties:
      created_at:
        type: string
      description:
        type: string
      field_type:
        type: string
      form_id:
        type: string
      id:
        type: string
      is_required:
        type: boolean
      key:
        type: string
      label:
        type: string
      options: {}
      placeholder:
        type: string
      section_id:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
      validation: {}
    type: object
  forms.FormResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      organization_id:
        type: string
      status:
        type: string
      template_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  forms.FormSectionWithFields:
    properties:
      fields:
        items:
          $ref: '#/definitions/forms.FormFieldDetail'
        type: array
      section:
        $ref: '#/definitions/forms.SectionDetailResponse'
    type: object
  forms.SectionDetailResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      form_id:
        type: string
      id:
        type: string
      sort_order:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  forms.UpdateFormDto:
    properties:
      description:
        maxLength: 1000
        type: string
      status:
        enum:
        - draft
        - active
        - archived
        type: string
      title:
        maxLength: 255
        minLength: 2
        type: string
    required:
    - title
    type: object
...
  utils.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  utils.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Preuvio.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Preuvio API
  version: "1.0"
paths:
...
  /api/forms:
    get:
      description: Retrieves all forms for the authenticated user's organization
      parameters:
      - description: Page (1-based, default 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Limit (default 20, max 50)
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Forms retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/forms.FormResponse'
                  type: array
                meta:
                  $ref: '#/definitions/utils.PaginationMeta'
              type: object
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List org forms (paginated)
      tags:
      - forms
    post:
      consumes:
      - application/json
      description: Creates a new empty form for the authenticated user's organization.
        Use POST /templates/{id}/clone to create a form from a template.
      parameters:
      - description: Form payload (scratch build - no template_id)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/forms.CreateFormDto'
      produces:
      - application/json
      responses:
        "201":
          description: Form created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/forms.FormResponse'
              type: object
        "400":
          description: Bad request or validation error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a scratch form
      tags:
      - forms
...
  /api/forms/{id}:
    delete:
      description: Permanently deletes a form
      parameters:
      - description: Form UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Form deleted successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Form not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a form
      tags:
      - forms
    get:
      description: Retrieves a form by ID with its sections and fields
      parameters:
      - description: Form UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Form retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/forms.FormResponse'
              type: object
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Form not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a form
      tags:
      - forms
    put:
      consumes:
      - application/json
      description: Updates an existing form's details
      parameters:
      - description: Form UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated form payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/forms.UpdateFormDto'
      produces:
      - application/json
      responses:
        "200":
          description: Form updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/forms.FormResponse'
              type: object
        "400":
          description: Bad request or validation error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Form not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a form
      tags:
      - forms
  /api/forms/{id}/detail:
    get:
      description: Retrieves a form with all its sections and nested fields in one
        call
      parameters:
      - description: Form UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Form detail retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/forms.FormDetailResponse'
              type: object
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Form not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get full form with sections and fields
      tags:
      - forms
...
swagger: "2.0"